Migrate log2 tests
